# Implemented faster algorithm for number of partitions