This implementation computes `npartitions(10**8)` in `2 s`, and `npartitions(10**9)` in `18 s` in my pc.
 The current sympy code takes `5 s` for `npartitions(10**6)`, `50 minutes` for `10**8` in my pc.
Mathematica 8 takes `23 s` for `n = 10**9` on a relatively very fast computer(maybe 5-6 times faster than mine).
I have verified the answers upto `10**11` with sage (which uses flint for this computation.)