@smichr 2016-03-28
I see this is trading speed for memory. Is storing `10**5` ints the same as storing bools? Would there be any advantage to using `_factor = defaultdict(int)` for which `_factor[i] -> 0` automatically without needing to store the value. Once accessed, however, the key is entered but if needed, keys that have zero values could be deleted to free up memory after some strategic point in the algorithm (if some exists).