@jtnydv25 2016-03-29
@smichr It is necessary to store `_factor`, and `_totient` for fast computation.
An alternative is using `factorint`, which'd be quite slow in comparison to this.
Also, I think `10**5` ints isn't a lot of memory.
You can easily store `10**8`. So, only about `0.1 %` of the memory available is used(that too only when `npartitions` is called for the first time.)