@jksuom 2016-03-29
I have tried to check this against Whiteman's paper and it looks mathematically ok to me. On the formal side, the long lines should be split.