@smichr 2016-03-30
touchups are [here](https://github.com/jtnydv25/sympy/pull/2)