@jksuom 2016-03-30
Looks good to me. (I personally like the formatting with space around ``%.``) Thank you for your work, @jtnydv25 ! Merging.